[projectfirma/#1238] Control display order of Taxonomy and Classification Items Refactored sort order control for generality. There is now an "IHaveASortOrder" Interface and Classification now implements this interface. The new drag-and-drop sorting control can be applied anywhere the IHaveASortOrder interface is used.
